pyproject.toml
[tool.poetry]
name = "influenceorbitmapper"
version = "0.1.0"
description = "Influence orbit visualization and route planning"
authors = ["Michael Verdegaal michaelverdegaal@hotmail.com"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
Quart = "^0.15.1"
ujson = "^4.2.0"
Jinja2 = "^3.0.2"
pandas = "^1.3.4"
numba = "^0.54.1"
pendulum = "^2.1.2"
libsass = "^0.21.0"
scikit-learn = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
